+++ Extra/Settings/Changelog.md
@@ -64,6 +64,7 @@ First released version!
 + Actual color pickers instead of 3 sliders
 + Update failed puzzles instantly from tab list
 + Instantly checkmark blaze puzzle after all blazes are done
++ box around wither doors on br
 ! fix secrets under head setting showing secrets over head with tenios style
 ! fix secret tick style showing ?/? for ironman players
 ! fix spirit leap menu not switching to boss tab in boss